Remove placeholder libraries and add requirements
