+++ requirements.txt
@@ -1,0 +1,11 @@
+numpy
+pandas
+scikit-learn
+rich
+matplotlib
+seaborn
+nltk
+textblob
+wordcloud
+torch
+transformers
