--- rich/__init__.py
@@ -1,2 +1,0 @@
-from .console import Console
-from .table import Table
