--- rich/console.py
@@ -1,3 +1,0 @@
-class Console:
-    def print(self, *args, **kwargs):
-        pass
