--- rich/table.py
@@ -1,7 +1,0 @@
-class Table:
-    def __init__(self, *args, **kwargs):
-        pass
-    def add_column(self, *args, **kwargs):
-        pass
-    def add_row(self, *args, **kwargs):
-        pass
