--- sklearn/__init__.py
@@ -1,1 +1,0 @@
-# minimal sklearn stub
