--- sklearn/cluster/__init__.py
@@ -1,5 +1,0 @@
-class KMeans:
-    def __init__(self, *args, **kwargs):
-        pass
-    def fit(self, *args, **kwargs):
-        pass
