--- sklearn/metrics/pairwise.py
@@ -1,2 +1,0 @@
-def cosine_similarity(X, Y=None):
-    return []
